requirements.txt
# Core dependencies for 5S Vision Checker
google-genai==1.24.0
Pillow==11.2.1
opencv-python==4.11.0.86
torch==2.7.1
torchvision==0.22.1
transformers==4.52.4
supervision==0.25.1
matplotlib==3.10.3
requests==2.32.4